refactor(Card): 불변으로 변경
