+++ src/main/java/blackjack/domain/gamer/Dealer.java
@@ -5,6 +5,7 @@ import blackjack.domain.card.Card;
 public class Dealer extends Gamer {
     private static final String NAME = "딜러";
     private static final int DEALER_STAND_CONDITION = 16;
+    private static final int INITIAL_OPEN_CARD_INDEX = 1;
 
     public Dealer() {
         super(NAME);
@@ -22,5 +23,9 @@ public class Dealer extends Gamer {
     @Override
     public boolean isNotBust() {
         return getScore() <= DEALER_STAND_CONDITION;
+    }
+
+    public Card getInitialOpenedCard() {
+        return getCards().get(INITIAL_OPEN_CARD_INDEX);
     }
 }
