+++ src/test/java/blackjack/domain/DealerTest.java
@@ -3,8 +3,8 @@ package blackjack.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.card.Card;
-import blackjack.domain.card.property.CardNumber;
-import blackjack.domain.card.property.CardShape;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardShape;
 import blackjack.domain.gamer.Dealer;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
